Week-05/animal-danger-detection/backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
Pillow==10.0.0
transformers
torch==2.5.1